Add Visual Studio project and solution files
The project doesn't build since Visual Studio wants to use its version
of TypeScript while React is using Babel.  The only way to get it to
build is to modify the import statements and have Visual Studio use its
own configuration.  This isn't necessary Visual Studio doesn't need to
build the project.  That is the job of `yarn build`.