Use invoke-handle as well as send-on (#76)
* Use invoke-handle instead of send-on

* Use MessageBoxOptions type

* Add send-on example

The startTimer and stopTimer functions have no interesting return value
so they use "send" instead of "invoke" in the renderer process and "on"
instead of "handle" in the main process.  The "main" channel is an
example of the use of "send" in the main process and "on" in the
renderer process.